third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-z-index-expected.txt


	






	





PASS successfullyParsed is true

TEST COMPLETE
PASS gFocusedDocument.activeElement.getAttribute("id") is "s22"
PASS gFocusedDocument.activeElement.getAttribute("id") is "p11"
PASS gFocusedDocument.activeElement.getAttribute("id") is "s23"
PASS gFocusedDocument.activeElement.getAttribute("id") is "p12"
PASS gFocusedDocument.activeElement.getAttribute("id") is "s24"
PASS gFocusedDocument.activeElement.getAttribute("id") is "s25"
PASS gFocusedDocument.activeElement.getAttribute("id") is "s24"
PASS gFocusedDocument.activeElement.getAttribute("id") is "p12"
PASS gFocusedDocument.activeElement.getAttribute("id") is "p22"
PASS gFocusedDocument.activeElement.getAttribute("id") is "p12"
PASS gFocusedDocument.activeElement.getAttribute("id") is "s24"
PASS gFocusedDocument.activeElement.getAttribute("id") is "s14"
PASS gFocusedDocument.activeElement.getAttribute("id") is "s13"
PASS gFocusedDocument.activeElement.getAttribute("id") is "s12"
PASS gFocusedDocument.activeElement.getAttribute("id") is "s22"
PASS gFocusedDocument.activeElement.getAttribute("id") is "p11"
PASS gFocusedDocument.activeElement.getAttribute("id") is "p21"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
This test is testing that we can prefer the elements with higher Z indexes.
